requirements.txt
anthropic
requests
beautifulsoup4
feedparser
python-dotenv
